Former-commit-id: fbce036034b963360622aaf096dc084faa6c04c3 [formerly 985e1ff6b6ea6487a3687fde67274a9c17706e14]
Former-commit-id: dff109d6a53a905f50abb487ed3575d4c55f3abc

+++ pyworkflow/em/packages/relion/viewer.py
@@ -249,10 +249,10 @@ Examples:
         
     def _formatFreq(self, value, pos):
         """ Format function for Matplotlib formatter. """
-        inv = 999
+        inv = 999.
         if value:
-            inv = int(1/value)
-        return "1/%d" % inv
+            inv = 1/value
+        return "1/%0.2f" % inv
 
     def _getGridSize(self, n=None):
         """ Figure out the layout of the plots given the number of references. """
